refactor: reorganize component structure for enterprise scalability
- Move reusable components from app/components to components/ directory
- Break down large components into smaller, focused modules
- Create organized folder structure: layout/, hero/, products/, marketing/, categories/, common/
- Split Header into SearchBar, UserMenu, CartIcon, Navigation
- Split Footer into FooterLinks, Newsletter
- Split HeroCarousel into HeroSlide, CarouselControls, types
- Split ProductCard into ProductImage, ProductInfo, PriceDisplay, AddToCartButton
- Split FlashSale into SaleTimer, SaleProducts
- Keep only page-specific components in app/components (NewYearSale, PromoBanners, ProductSection)
- Update all imports to use @/components/ paths
- Improve maintainability and team collaboration for 1M+ user scale
