--- Frontend/app/components/Footer.tsx
@@ -1,176 +1,0 @@
-import Link from "next/link"
-import { Facebook, Twitter, Instagram, Youtube, Mail, Phone, MapPin } from "lucide-react"
-
-export default function Footer() {
-  return (
-    <footer className="bg-gray-900 text-white">
-      <div className="container mx-auto px-4 py-12">
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8">
-          {/* Company Info */}
-          <div>
-            <div className="flex items-center space-x-2 mb-4">
-              <div className="w-8 h-8 bg-orange-500 rounded-lg flex items-center justify-center">
-                <span className="text-white font-bold">i</span>
-              </div>
-              <span className="text-2xl font-bold">iShop</span>
-            </div>
-            <p className="text-gray-400 mb-4">
-              Your trusted online shopping destination for electronics, fashion, and more. Quality products at
-              unbeatable prices.
-            </p>
-            <div className="space-y-2">
-              <div className="flex items-center space-x-2 text-gray-400">
-                <MapPin className="w-4 h-4" />
-                <span>123 Shopping Street, Mumbai, India</span>
-              </div>
-              <div className="flex items-center space-x-2 text-gray-400">
-                <Phone className="w-4 h-4" />
-                <span>+91 9876543210</span>
-              </div>
-              <div className="flex items-center space-x-2 text-gray-400">
-                <Mail className="w-4 h-4" />
-                <span>support@ishop.com</span>
-              </div>
-            </div>
-          </div>
-
-          {/* Quick Links */}
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Quick Links</h3>
-            <ul className="space-y-2">
-              <li>
-                <Link href="/about" className="text-gray-400 hover:text-white transition-colors">
-                  About Us
-                </Link>
-              </li>
-              <li>
-                <Link href="/contact" className="text-gray-400 hover:text-white transition-colors">
-                  Contact Us
-                </Link>
-              </li>
-              <li>
-                <Link href="/careers" className="text-gray-400 hover:text-white transition-colors">
-                  Careers
-                </Link>
-              </li>
-              <li>
-                <Link href="/blog" className="text-gray-400 hover:text-white transition-colors">
-                  Blog
-                </Link>
-              </li>
-              <li>
-                <Link href="/press" className="text-gray-400 hover:text-white transition-colors">
-                  Press
-                </Link>
-              </li>
-              <li>
-                <Link href="/sitemap" className="text-gray-400 hover:text-white transition-colors">
-                  Sitemap
-                </Link>
-              </li>
-            </ul>
-          </div>
-
-          {/* Customer Service */}
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Customer Service</h3>
-            <ul className="space-y-2">
-              <li>
-                <Link href="/help" className="text-gray-400 hover:text-white transition-colors">
-                  Help Center
-                </Link>
-              </li>
-              <li>
-                <Link href="/returns" className="text-gray-400 hover:text-white transition-colors">
-                  Returns & Refunds
-                </Link>
-              </li>
-              <li>
-                <Link href="/shipping" className="text-gray-400 hover:text-white transition-colors">
-                  Shipping Info
-                </Link>
-              </li>
-              <li>
-                <Link href="/track-order" className="text-gray-400 hover:text-white transition-colors">
-                  Track Your Order
-                </Link>
-              </li>
-              <li>
-                <Link href="/warranty" className="text-gray-400 hover:text-white transition-colors">
-                  Warranty
-                </Link>
-              </li>
-              <li>
-                <Link href="/bulk-orders" className="text-gray-400 hover:text-white transition-colors">
-                  Bulk Orders
-                </Link>
-              </li>
-            </ul>
-          </div>
-
-          {/* Newsletter & Social */}
-          <div>
-            <h3 className="text-lg font-semibold mb-4">Stay Connected</h3>
-            <p className="text-gray-400 mb-4">Follow us on social media for latest updates and exclusive offers.</p>
-            <div className="flex space-x-4 mb-6">
-              <Link href="#" className="text-gray-400 hover:text-white transition-colors">
-                <Facebook className="w-6 h-6" />
-              </Link>
-              <Link href="#" className="text-gray-400 hover:text-white transition-colors">
-                <Twitter className="w-6 h-6" />
-              </Link>
-              <Link href="#" className="text-gray-400 hover:text-white transition-colors">
-                <Instagram className="w-6 h-6" />
-              </Link>
-              <Link href="#" className="text-gray-400 hover:text-white transition-colors">
-                <Youtube className="w-6 h-6" />
-              </Link>
-            </div>
-            <div>
-              <h4 className="font-medium mb-2">Download Our App</h4>
-              <div className="flex flex-col space-y-2">
-                <Link href="#" className="inline-block">
-                  <div className="bg-gray-800 px-4 py-2 rounded-lg border border-gray-700 hover:bg-gray-700 transition-colors">
-                    <div className="text-xs text-gray-400">Download on the</div>
-                    <div className="text-sm font-medium">App Store</div>
-                  </div>
-                </Link>
-                <Link href="#" className="inline-block">
-                  <div className="bg-gray-800 px-4 py-2 rounded-lg border border-gray-700 hover:bg-gray-700 transition-colors">
-                    <div className="text-xs text-gray-400">Get it on</div>
-                    <div className="text-sm font-medium">Google Play</div>
-                  </div>
-                </Link>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Bottom Bar */}
-        <div className="border-t border-gray-800 mt-12 pt-8">
-          <div className="flex flex-col md:flex-row justify-between items-center">
-            <div className="text-gray-400 text-sm mb-4 md:mb-0">
-              © 2024 iShop. All rights reserved. |
-              <Link href="/privacy" className="hover:text-white ml-1">
-                Privacy Policy
-              </Link>{" "}
-              |
-              <Link href="/terms" className="hover:text-white ml-1">
-                Terms of Service
-              </Link>
-            </div>
-            <div className="flex items-center space-x-4">
-              <span className="text-gray-400 text-sm">We Accept:</span>
-              <div className="flex space-x-2">
-                <div className="bg-white px-2 py-1 rounded text-xs font-bold text-gray-800">bKash</div>
-                <div className="bg-white px-2 py-1 rounded text-xs font-bold text-gray-800">Nagad</div>
-                <div className="bg-white px-2 py-1 rounded text-xs font-bold text-gray-800">VISA</div>
-                <div className="bg-white px-2 py-1 rounded text-xs font-bold text-gray-800">Rocket</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </footer>
-  )
-}
