--- Frontend/app/components/Newsletter.tsx
@@ -1,60 +1,0 @@
-"use client"
-
-import type React from "react"
-
-import { useState } from "react"
-import { Mail, CheckCircle } from "lucide-react"
-
-export default function Newsletter() {
-  const [email, setEmail] = useState("")
-  const [isSubscribed, setIsSubscribed] = useState(false)
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    if (email) {
-      setIsSubscribed(true)
-      setEmail("")
-      setTimeout(() => setIsSubscribed(false), 3000)
-    }
-  }
-
-  return (
-    <section className="py-16 bg-gradient-to-r from-orange-50 to-red-50">
-      <div className="container mx-auto px-4 text-center">
-        <div className="max-w-2xl mx-auto">
-          <Mail className="w-16 h-16 text-orange-500 mx-auto mb-6" />
-          <h2 className="text-3xl md:text-4xl font-bold text-gray-800 mb-4">Subscribe to Newsletter</h2>
-          <p className="text-gray-600 mb-8 text-lg">
-            Get latest updates and offers directly in your inbox. Be the first to know about our exclusive deals!
-          </p>
-
-          {isSubscribed ? (
-            <div className="flex items-center justify-center space-x-2 text-green-600">
-              <CheckCircle className="w-6 h-6" />
-              <span className="text-lg font-medium">Thank you for subscribing!</span>
-            </div>
-          ) : (
-            <form onSubmit={handleSubmit} className="flex flex-col sm:flex-row gap-4 max-w-md mx-auto">
-              <input
-                type="email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                placeholder="Enter your email address"
-                className="flex-1 px-4 py-3 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-orange-500 focus:border-transparent"
-                required
-              />
-              <button
-                type="submit"
-                className="bg-orange-500 text-white px-8 py-3 rounded-lg font-semibold hover:bg-orange-600 transition-colors whitespace-nowrap"
-              >
-                Subscribe
-              </button>
-            </form>
-          )}
-
-          <p className="text-sm text-gray-500 mt-4">We respect your privacy. Unsubscribe at any time.</p>
-        </div>
-      </div>
-    </section>
-  )
-}
