+++ Frontend/components/common/ServiceFeatures/index.tsx
@@ -54,4 +54,4 @@ export default function ServiceFeatures() {
       </div>
     </section>
   )
-}
+}
