+++ Frontend/components/hero/HeroCarousel/CarouselControls.tsx
@@ -1,0 +1,53 @@
+"use client"
+
+import { ChevronLeft, ChevronRight } from "lucide-react"
+
+interface CarouselControlsProps {
+  onPrevious: () => void
+  onNext: () => void
+  totalSlides: number
+  currentSlide: number
+  onSlideChange: (index: number) => void
+}
+
+export default function CarouselControls({
+  onPrevious,
+  onNext,
+  totalSlides,
+  currentSlide,
+  onSlideChange,
+}: CarouselControlsProps) {
+  return (
+    <>
+      {/* Navigation Arrows */}
+      <button
+        onClick={onPrevious}
+        className="absolute left-4 top-1/2 -translate-y-1/2 bg-white/20 hover:bg-white/30 text-white p-2 rounded-full transition-colors"
+        aria-label="Previous slide"
+      >
+        <ChevronLeft className="w-6 h-6" />
+      </button>
+      <button
+        onClick={onNext}
+        className="absolute right-4 top-1/2 -translate-y-1/2 bg-white/20 hover:bg-white/30 text-white p-2 rounded-full transition-colors"
+        aria-label="Next slide"
+      >
+        <ChevronRight className="w-6 h-6" />
+      </button>
+
+      {/* Indicators */}
+      <div className="absolute bottom-4 left-1/2 -translate-x-1/2 flex space-x-2">
+        {Array.from({ length: totalSlides }, (_, index) => (
+          <button
+            key={index}
+            onClick={() => onSlideChange(index)}
+            className={`w-3 h-3 rounded-full transition-colors ${
+              index === currentSlide ? "bg-white" : "bg-white/50"
+            }`}
+            aria-label={`Go to slide ${index + 1}`}
+          />
+        ))}
+      </div>
+    </>
+  )
+}
