+++ Frontend/components/products/ProductGrid/ProductGridItem.tsx
@@ -1,0 +1,25 @@
+import ProductCard from "../ProductCard"
+
+interface Product {
+  id: string
+  name: string
+  price: number
+  originalPrice?: number | null
+  discount?: number | null
+  rating: number
+  reviews: number
+  image: string
+  badge?: string
+}
+
+interface ProductGridItemProps {
+  product: Product
+}
+
+export default function ProductGridItem({ product }: ProductGridItemProps) {
+  return (
+    <div className="flex-shrink-0 w-64">
+      <ProductCard product={product} />
+    </div>
+  )
+}
